refactor(tests): replace shallow and mount in table components
