+++ src/components/Table/Table.js
@@ -13,12 +13,10 @@
  * limitations under the License.
  */
 
-/** @jsx jsx */
-
-import { Component } from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { jsx, css } from '@emotion/core';
+import { css } from '@emotion/core';
 import { isNil, throttle } from 'lodash/fp';
 
 import TableHead from './components/TableHead';
