+++ src/components/Table/Table.spec.js
@@ -14,174 +14,149 @@
  */
 
 import React from 'react';
-import { shuffle } from 'lodash/fp';
 
 import Table from './Table';
-import { ASCENDING, DESCENDING } from './constants';
+import { ASCENDING } from './constants';
 
-const headers = [{ children: 'Foo', sortable: true }, 'Bar', 'Baz'];
-const items = [['1', '2', '3'], ['1', '2', '3'], ['1', '2', '3']];
+const headers = [
+  { children: 'Letters', sortable: true },
+  { children: 'Numbers', sortable: true },
+  'Words'
+];
+const rows = [['b', '3', 'Foo'], ['a', '1', 'Bar'], ['c', '2', 'Baz']];
+const rowLength = rows[0].length;
 
 describe('Table', () => {
   beforeEach(jest.clearAllMocks);
 
   describe('Style tests', () => {
     it('should render with default styles', () => {
-      const actual = create(<Table headers={headers} rows={items} />);
+      const actual = create(<Table headers={headers} rows={rows} />);
       expect(actual).toMatchSnapshot();
     });
 
     it('should render without the table shadow', () => {
-      const actual = create(<Table headers={headers} rows={items} noShadow />);
+      const actual = create(<Table headers={headers} rows={rows} noShadow />);
       expect(actual).toMatchSnapshot();
     });
 
     it('should render with rowHeader styles', () => {
-      const actual = create(
-        <Table rowHeaders headers={headers} rows={items} />
-      );
+      const actual = create(<Table rowHeaders headers={headers} rows={rows} />);
       expect(actual).toMatchSnapshot();
     });
 
     it('should render a collapsed table', () => {
       const actual = create(
-        <Table headers={headers} rows={items} borderCollapsed />
+        <Table headers={headers} rows={rows} borderCollapsed />
       );
       expect(actual).toMatchSnapshot();
     });
 
     it('should render a condensed table', () => {
-      const actual = create(<Table headers={headers} rows={items} condensed />);
+      const actual = create(<Table headers={headers} rows={rows} condensed />);
       expect(actual).toMatchSnapshot();
     });
 
     it('should render a scrollable table', () => {
       const actual = create(
-        <Table headers={headers} rows={items} scrollable rowHeaders={false} />
+        <Table headers={headers} rows={rows} scrollable rowHeaders={false} />
       );
       expect(actual).toMatchSnapshot();
     });
 
     it('should not render a scrollable table if the rowHeaders prop is true', () => {
-      const actual = create(<Table headers={headers} rows={items} />);
+      const actual = create(<Table headers={headers} rows={rows} />);
       expect(actual).toMatchSnapshot();
+    });
+  });
+
+  describe('Interaction tests', () => {
+    it('should call the row click callback', () => {
+      const onRowClickMock = jest.fn();
+      const index = 0;
+      const { getAllByTestId } = render(
+        <Table onRowClick={onRowClickMock} headers={headers} rows={rows} />
+      );
+
+      act(() => {
+        fireEvent.click(getAllByTestId('table-row')[0]);
+      });
+
+      expect(onRowClickMock).toHaveBeenCalledTimes(1);
+      expect(onRowClickMock).toHaveBeenCalledWith(index);
+    });
+
+    describe('sorting', () => {
+      it('should sort a column in ascending order', () => {
+        const { getAllByTestId } = render(
+          <Table rows={rows} headers={headers} />
+        );
+
+        const letterHeaderEl = getAllByTestId('table-header')[0];
+        const cellEls = getAllByTestId('table-cell');
+
+        act(() => {
+          fireEvent.click(letterHeaderEl);
+        });
+
+        const sortedRow = ['a', 'b', 'c'];
+
+        rows.forEach((row, index) => {
+          // There's a hidden header cell that we need to skip with +1.
+          const cellIndex = rowLength * index + 1;
+          expect(cellEls[cellIndex]).toHaveTextContent(sortedRow[index]);
+        });
+      });
+
+      it('should sort a column in descending order', () => {
+        const { getAllByTestId } = render(
+          <Table rows={rows} headers={headers} />
+        );
+
+        const letterHeaderEl = getAllByTestId('table-header')[0];
+        const cellEls = getAllByTestId('table-cell');
+
+        act(() => {
+          fireEvent.click(letterHeaderEl);
+        });
+        act(() => {
+          fireEvent.click(letterHeaderEl);
+        });
+
+        const sortedRow = ['c', 'b', 'a'];
+
+        rows.forEach((row, index) => {
+          // There's a hidden header cell that we need to skip with +1.
+          const cellIndex = rowLength * index + 1;
+          expect(cellEls[cellIndex]).toHaveTextContent(sortedRow[index]);
+        });
+      });
+
+      it('should call a custom sort callback', () => {
+        const onSortByMock = jest.fn();
+        const index = 0;
+        const nextDirection = ASCENDING;
+        const { getAllByTestId } = render(
+          <Table onSortBy={onSortByMock} headers={headers} rows={rows} />
+        );
+
+        act(() => {
+          fireEvent.click(getAllByTestId('table-header')[0]);
+        });
+
+        expect(onSortByMock).toHaveBeenCalledTimes(1);
+        expect(onSortByMock).toHaveBeenCalledWith(index, nextDirection, rows);
+      });
     });
   });
 
   describe('Accessibility tests', () => {
     it('should meet accessibility guidelines', async () => {
       const wrapper = renderToHtml(
-        <Table rowHeaders headers={headers} rows={items} />
+        <Table rowHeaders headers={headers} rows={rows} />
       );
       const actual = await axe(wrapper);
       expect(actual).toHaveNoViolations();
-    });
-  });
-
-  describe('onSortEnter()', () => {
-    it('should set sortHover to the provided index', () => {
-      const wrapper = shallow(<Table />)
-        .find('Table')
-        .instance();
-
-      const index = 0;
-      const expected = index;
-
-      wrapper.onSortEnter(index);
-
-      const actual = wrapper.state.sortHover;
-
-      expect(actual).toBe(expected);
-    });
-  });
-
-  describe('onSortLeave()', () => {
-    it('should set sortHover to null', () => {
-      const wrapper = shallow(<Table />)
-        .find('Table')
-        .instance();
-
-      wrapper.setState({
-        sortHover: 0
-      });
-      wrapper.onSortLeave();
-
-      const actual = wrapper.state.sortHover;
-
-      expect(actual).toBeNull();
-    });
-  });
-
-  describe('onSortBy()', () => {
-    describe('custom onSortBy', () => {
-      it('should call the provided onSortBy instead of defaultSortBy with index, nextDirection and rows', () => {
-        const row = ['a', 'b', 'c', 'd', 'e'];
-        const rows = [row];
-        const shuffledRow = shuffle(row);
-        const expected = [shuffledRow];
-        const mock = jest.fn().mockReturnValue(expected);
-        const index = 0;
-        const nextDirection = ASCENDING;
-        const wrapper = shallow(<Table onSortBy={mock} rows={rows} />)
-          .find('Table')
-          .instance();
-
-        wrapper.onSortBy(index);
-
-        expect(mock).toHaveBeenCalledWith(index, nextDirection, rows);
-        expect(wrapper.getSortedRows()).toEqual(expected);
-      });
-    });
-  });
-
-  describe('updateSort()', () => {
-    it('should update the state with sortedRow, nextDirection and nextDirection', () => {
-      const wrapper = shallow(<Table />)
-        .find('Table')
-        .instance();
-      const index = 0;
-      const nextDirection = ASCENDING;
-
-      wrapper.updateSort(index, nextDirection);
-
-      expect(wrapper.state.sortedRow).toBe(index);
-      expect(wrapper.state.sortDirection).toBe(nextDirection);
-    });
-  });
-
-  describe('defaultSortBy()', () => {
-    describe('next direction ASCENDING', () => {
-      it('should sort the provided row by ascending order', () => {
-        const wrapper = shallow(<Table />)
-          .find('Table')
-          .instance();
-        const index = 0;
-        const nextDirection = ASCENDING;
-        const a = ['Foo'];
-        const b = ['Bar'];
-        const rows = [a, b];
-        const expected = [b, a];
-        const actual = wrapper.defaultSortBy(index, nextDirection, rows);
-
-        expect(actual).toEqual(expected);
-      });
-    });
-
-    describe('next direction DESCENDING', () => {
-      it('should sort the provided row by descending order', () => {
-        const wrapper = shallow(<Table />)
-          .find('Table')
-          .instance();
-        const index = 0;
-        const nextDirection = DESCENDING;
-        const a = ['Bar'];
-        const b = ['Foo'];
-        const rows = [a, b];
-        const expected = [b, a];
-        const actual = wrapper.defaultSortBy(index, nextDirection, rows);
-
-        expect(actual).toEqual(expected);
-      });
     });
   });
 });
