+++ src/components/Table/components/SortArrow/index.js
@@ -13,64 +13,6 @@
  * limitations under the License.
  */
 
-import React, { Fragment } from 'react';
-import PropTypes from 'prop-types';
-import styled from '@emotion/styled';
-import { css } from '@emotion/core';
-import { ReactComponent as ArrowIcon } from './arrow.svg';
-
-import SvgButton from '../../../SvgButton';
-import { ASCENDING, DESCENDING } from '../../constants';
-
-const baseStyles = ({ theme }) => css`
-  display: flex;
-  flex-direction: column;
-  height: 10px;
-  left: ${theme.spacings.kilo};
-  opacity: 0;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  transition: opacity ${theme.transitions.default};
-  width: 5px;
-`;
-
-const condensedStyles = ({ condensed, theme }) =>
-  condensed &&
-  css`
-    left: ${theme.spacings.byte};
-  `;
-
-const StyledWrapper = styled(SvgButton)`
-  ${baseStyles};
-  ${condensedStyles};
-`;
-
-const DownArrow = styled(ArrowIcon)`
-  margin-top: 2px;
-  transform: rotate(180deg);
-`;
-
-/**
- * @private Arrow component for TableHeader sorting
- */
-const SortArrow = ({ direction, condensed }) => (
-  <StyledWrapper condensed={condensed}>
-    <Fragment>
-      {direction !== DESCENDING && <ArrowIcon />}
-      {direction !== ASCENDING && <DownArrow />}
-    </Fragment>
-  </StyledWrapper>
-);
-
-SortArrow.propTypes = {
-  direction: PropTypes.oneOf([ASCENDING, DESCENDING]),
-  condensed: PropTypes.bool
-};
-
-SortArrow.defaultProps = {
-  direction: null,
-  condensed: false
-};
+import SortArrow from './SortArrow';
 
 export default SortArrow;
