+++ src/components/Table/components/TableBody/index.js
@@ -13,106 +13,6 @@
  * limitations under the License.
  */
 
-import React, { Fragment } from 'react';
-import PropTypes from 'prop-types';
-
-import TableRow from '../TableRow';
-import TableHeader from '../TableHeader';
-import TableCell from '../TableCell';
-import {
-  mapRowProps,
-  mapCellProps,
-  getCellChildren,
-  RowPropType
-} from '../../utils';
-import { TR_KEY_PREFIX, TD_KEY_PREFIX } from '../../constants';
-
-const getRowKey = index => `${TR_KEY_PREFIX}-${index}`;
-const getCellKey = (rowIndex, cellIndex) =>
-  `${TD_KEY_PREFIX}-${rowIndex}-${cellIndex}`;
-
-const TableBody = ({ rows, condensed, rowHeaders, sortHover, onRowClick }) => (
-  <tbody>
-    {rows.map((row, rowIndex) => {
-      const { cells, ...props } = mapRowProps(row);
-      return (
-        <TableRow
-          key={getRowKey(rowIndex)}
-          onClick={onRowClick && (() => onRowClick(rowIndex))}
-          {...props}
-        >
-          {cells.map((cell, cellIndex) =>
-            rowHeaders && cellIndex === 0 ? (
-              <Fragment key={getCellKey(rowIndex, cellIndex)}>
-                <TableHeader
-                  fixed
-                  condensed={condensed}
-                  scope={TableHeader.ROW}
-                  isHovered={sortHover === cellIndex}
-                  {...mapCellProps(cell)}
-                />
-                <TableCell
-                  role="presentation"
-                  condensed={condensed}
-                  aria-hidden="true"
-                >
-                  {getCellChildren(cell)}
-                </TableCell>
-              </Fragment>
-            ) : (
-              <TableCell
-                key={getCellKey(rowIndex, cellIndex)}
-                condensed={condensed}
-                isHovered={sortHover === cellIndex}
-                {...mapCellProps(cell)}
-              />
-            )
-          )}
-        </TableRow>
-      );
-    })}
-  </tbody>
-);
-
-/**
- * @private TableHead for the Table component. The Table handles rendering it
- */
-TableBody.propTypes = {
-  /**
-   *(An array of rows or object with children) containing cells for the table. The Cell can be a string
-   * or an object with options described on TableHeader component
-   */
-  rows: PropTypes.arrayOf(
-    PropTypes.oneOfType([
-      PropTypes.shape({ cells: PropTypes.arrayOf(RowPropType) }),
-      PropTypes.arrayOf(RowPropType)
-    ])
-  ),
-  /**
-   * Enables/disables sticky columns on mobile
-   */
-  rowHeaders: PropTypes.bool,
-  /**
-   * @private The current hovered sort cell index
-   * Handled internally
-   */
-  sortHover: PropTypes.number,
-  /**
-   * @private Adds condensed styles to the table.
-   * Handled internally
-   */
-  condensed: PropTypes.bool,
-  /**
-   * onClick handler
-   */
-  onRowClick: PropTypes.func
-};
-
-TableBody.defaultProps = {
-  rows: [],
-  rowHeaders: false,
-  sortHover: null,
-  onRowClick: null
-};
+import TableBody from './TableBody';
 
 export default TableBody;
