+++ src/components/Table/components/TableCell/TableCell.spec.js
